Check for the right indent in configure.in and let the user specify one
Check for the right indent in configure.in and let the user specify
one there. Also fixed the KR style by adding -psl to the command
line. Now ORBit compiles with --enable-indent=kr (It isn't K&R any
more though...., but almost [the type of a function is now on a
seperate line])